pyproject.toml
[project]
name = "autos"
version = "0.1.0"
description = "Auto-Shorts CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "scenedetect[opencv]>=0.6.7.1",
    "typer>=0.21.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[project.scripts]
auto = "autos.cli:main"

# Standard Python packaging section
[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"



# Tell uv that this project should install itself
[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]
